netlog - util -> to psql
