+++ internal/netlog/psql_api.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"strings"
 	"time"
 
 	"github.com/jackc/pgx/v4/pgxpool"
@@ -211,4 +212,34 @@ func (api *PsqlApi) GetVisitsPage(keywords []string, field string, source string
 	}
 
 	return visits, nil
+}
+
+// getQueryWhereCondition will make a SQL WHERE condition
+// keywords starting with "-" will be filtered out with `url NOT LIKE ...`
+// column - the name of the column to which the "like" is applied for
+// source - the source of the netlog visit
+func getQueryWhereCondition(column, source string, keywords []string) string {
+	var sbQueryLike strings.Builder
+	if len(keywords) > 0 {
+		sbQueryLike.WriteString("WHERE ")
+		for i, word := range keywords {
+			if strings.HasPrefix(word, "-") {
+				word = strings.TrimPrefix(word, "-")
+				sbQueryLike.WriteString(fmt.Sprintf("%s NOT LIKE '%%%s%%' ", column, word))
+			} else {
+				sbQueryLike.WriteString(fmt.Sprintf("%s LIKE '%%%s%%' ", column, word))
+			}
+			if i < len(keywords)-1 {
+				sbQueryLike.WriteString("AND ")
+			}
+		}
+	}
+
+	if source != "all" && len(keywords) == 0 {
+		sbQueryLike.WriteString(fmt.Sprintf("WHERE source = '%s'", source))
+	} else if source != "all" {
+		sbQueryLike.WriteString(fmt.Sprintf("AND source = '%s'", source))
+	}
+
+	return sbQueryLike.String()
 }
